polish collision graph
