+++ Assets/Scripts/Graph/BasicNode.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class BasicNode : GraphNode
+{
+    public BasicNode(Vector3 normal, Vector3 realWorldPosition, Graph graph) : base(normal, realWorldPosition, graph) { }
+
+}
