+++ Assets/Scripts/Graph/BasicNode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a47a167c020bdb541acdc71f6e849293
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
