+++ Assets/Scripts/Graph/VoidNode.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class VoidNode : GraphNode
+{
+    public VoidNode(Vector3 normal, Vector3 realWorldPosition, Graph graph) : base(normal, realWorldPosition, graph) { }
+}
