+++ Assets/Scripts/Graph/VoidNode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fc8ef9824b93e7b4cb1f884f65a4e62c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
